Added Delete and first try at patch.

+++ phoodchef/Models/DTOs/RecipeDto.cs
@@ -9,7 +9,7 @@ namespace phoodchef.Models.DTOs
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int? CookTime { get; set; }
+        public decimal? CookTime { get; set; }
         public string CookUnit { get; set; }
         public string Instructions { get; set; }
         public string Yield { get; set; }
